+++ EngLang/StupidEngLangParser.cs
@@ -62,7 +62,7 @@ public class StupidEngLangParser : IEngLangParser
             var defaultValueExpression = defaultValue is null ? null
                 : defaultValue.StartsWith("\"") ? (Expression?)new StringLiteralExpression(defaultValue, default)
                 : defaultValue.StartsWith("$") ? (Expression?)new StringLiteralExpression(defaultValue, default)
-                : new IntLiteralExpression(int.Parse(defaultValue), default);
+                : new IntLiteralExpression(long.Parse(defaultValue), default);
             Statement declStat = new VariableDeclarationStatement(
                 new VariableDeclaration(variableName, new TypeIdentifierReference(typeName, false, default), defaultValueExpression, default),
                 default);
